hw1/p2_code/dataset.py
import os
from PIL import Image
from torch.utils.data import Dataset
# "ConcatDataset" and "Subset" are possibly useful when doing semi-supervised learning.

class P2_mini(Dataset): # for self-supervised learning
    def __init__(self, path, tfm = None, file_names = None):
        super(P2_mini).__init__()
        self.transform = tfm
        self.path = path
        self.file_names = sorted([os.path.join(path,x) for x in os.listdir(path) if x.endswith(".jpg")])
        if file_names != None:
            self.file_names = file_names
            print(f"One {path} sample",self.file_names)
        else:
            print(f"One {path} sample",self.file_names[0])
        
    def __len__(self):
        return len(self.file_names)
  
    def __getitem__(self,idx):
        fname = self.file_names[idx]
        im = Image.open(fname)
        im = self.transform(im)
            
        return im
    
class P2_office(Dataset):
    def __init__(self, path, tfm = None, file_names = None):
        super(P2_office).__init__()
        self.transform = tfm
        self.path = path
        self.file_names = sorted([os.path.join(path,x) for x in os.listdir(path) if x.endswith(".jpg")])
        if file_names != None:
            self.file_names = file_names
            print(f"One {path} sample",self.file_names)
        else:
            print(f"One {path} sample",self.file_names[0])
  
    def __len__(self):
        return len(self.file_names)
  
    def __getitem__(self,idx):
        fname = self.file_names[idx]
        im = Image.open(fname)
        im = self.transform(im)
        try:
            label = int(fname.split("/")[-1].split("_")[0])
        except:
            label = -100
            
        return im, label